pyproject.toml
[tool.poetry]
name = "veriflow"
version = "0.1.0"
description = ""
authors = ["Faried Abu Zaid <fariedaz@gmail.com>"]
readme = "README.md"
packages = [{include = "veriflow"}]

[tool.poetry.dependencies]
python = "^3.11"
pyro-ppl = "^1.8.5"
scikit-learn = "^1.3.0"
ray = "^2.5.1"
pandas = "^2.0.3"
pyarrow = "^12.0.1"
click = "^8.1.6"
tensorboardx = "^2.6.1"
idx2numpy = "^1.2.3"
numpy = "^1.25.1"
torchvision = "^0.15.2"
torch = "^2.0.1"
torchaudio = "^2.0.2"
fsspec = "^2023.6.0"
rise = "^5.7.1"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
matplotlib = "^3.7.2"
jupyter-contrib-nbextensions = "^0.7.0"
jupyter = "^1.0.0"
notebook = "6.4.8"
spellchecker = "^0.4"
tensorboard = "^2.14.0"
tox = "^4.11.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
testpaths = [
    "tests",
]